Daniel develop

+++ miniMapPrototype/src/net/devilishro/minimap/network/Opcode.java
@@ -1,5 +1,5 @@
 package net.devilishro.minimap.network;
 
 public interface Opcode {
-    public int getValue();
+    public short getValue();
 }
